refactor(user): 유저 인증 정보 dto 네이밍 변경
